CORE-209 replaced extension methods
