+++ frontend/apps/quarano-frontend/src/app/modules/tenant-admin/client/anomaly/anomaly.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { ActionDto } from '../../../../models/case-action';
+import { DateFunctions } from '@quarano-frontend/shared/util';
 
 @Component({
   selector: 'qro-client-action-anomaly',
@@ -14,6 +15,6 @@ export class AnomalyComponent implements OnInit {
   }
 
   getAnomalyDate(date: string): string {
-    return new Date(date).toCustomLocaleDateString();
+    return DateFunctions.toCustomLocaleDateString(new Date(date));
   }
 }
