+++ frontend/apps/quarano-frontend/tsconfig.app.json
@@ -4,6 +4,5 @@
     "outDir": "../../dist/out-tsc",
     "types": []
   },
-  "files": ["src/main.ts", "src/polyfills.ts"],
-  "include": ["**/*.d.ts"]
+  "files": ["src/main.ts", "src/polyfills.ts"]
 }
