+++ frontend/libs/health-department/contact-cases/case-list/src/lib/case-list/case-list.component.html
@@ -3,15 +3,6 @@
     <div style="display: flex;">
       <button
         mat-raised-button
-        data-cy="new-case-button"
-        color="primary"
-        [routerLink]="['/tenant-admin/client', 'index']"
-      >
-        <mat-icon>add</mat-icon>Neuen Indexfall anlegen
-      </button>
-      <button
-        class="ml-3"
-        mat-raised-button
         color="primary"
         [routerLink]="['/tenant-admin/client', 'contact']"
       >
