+++ frontend/libs/health-department/index-cases/case-list/src/lib/case-list/case-list.component.html
@@ -9,14 +9,6 @@
       >
         <mat-icon>add</mat-icon>Neuen Indexfall anlegen
       </button>
-      <button
-        class="ml-3"
-        mat-raised-button
-        color="primary"
-        [routerLink]="['/tenant-admin/client', 'contact']"
-      >
-        <mat-icon>add</mat-icon>Neuen Kontaktfall anlegen
-      </button>
     </div>
 
     <button data-cy="export-cases-button" mat-button>
