+++ frontend/libs/shared/util/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/shared-util.module';
 export * from './lib/constants';
-export * from './lib/group-by';
+export * from './lib/array-functions';
 export * from './lib/date-functions';
