--- frontend/libs/shared/util/src/lib/extensions/array-extensions.d.ts
@@ -1,7 +1,0 @@
-interface Array<T> {
-  remove(elem: T): Array<T>;
-}
-
-Array.prototype.remove = function <T>(elem: T): T[] {
-  return this.filter((e: T) => e !== elem);
-};
